+++ CCMenuTests/Classes/CCMConnectionTest.m
@@ -15,7 +15,7 @@
     connectionMock = [OCMockObject partialMockForObject:connection];
 }
 
-- (id)responseMockWithStatusCode:(int)statusCode
+- (id)responseMockWithStatusCode:(NSInteger)statusCode
 {
     id responseMock = [OCMockObject mockForClass:[NSHTTPURLResponse class]];
     [[[responseMock stub] andReturnValue:OCMOCK_VALUE(statusCode)] statusCode];
